proffinder/requirements.txt
streamlit>=1.28.0
httpx>=0.25.0
pydantic>=2.0.0
tenacity>=8.2.0
beautifulsoup4>=4.12.0
selectolax>=0.3.16
diskcache>=5.6.3
pandas>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
typing-extensions>=4.7.0
aiofiles>=23.0.0